+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Paragraphs with attributes being removed
+- crash when HTML tags contained special regex characters ('$' or '\\')
 
 ## [1.0-beta.9](https://github.com/wordpress-mobile/AztecEditor-Android/releases/tag/v1.0-beta.9) - 2017-09-25
 ### Fixed